design : 💄 스타일 파일 분리
